requirements.txt
streamlit
mysql-connector-python
bcrypt
mysqlclient
apache-airflow
apache-airflow-providers-mysql
python-dotenv
requests
tabulate
pillow
pandas